requirements.txt
python-telegram-bot==20.7
playwright==1.42.0
playwright-stealth==1.0.6
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.1
aiohttp==3.9.3
lxml==5.1.0
requests==2.31.0
cloudscraper==1.2.71
